requirements.txt
django_url_filter==0.3.4
Django==2.0.1
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.3.2
dj-database-url==0.4.2
django-cors-headers==2.1.0
